Corrected id type

+++ DC1AP/Items/Attachment.cs
@@ -4,7 +4,7 @@ namespace DC1AP.Items
     internal class Attachment
     {
         public string Name;
-        public int ItemID;
+        public short ItemID;
         public int[] Values;
         public int[] ValueOffsets;
     }
